dragonboat: NodeHost methods called after Stop() will return ErrClosed

+++ nodehost.go
@@ -367,7 +367,9 @@ func (nh *NodeHost) RaftAddress() string {
 // Stop stops all Raft nodes managed by the NodeHost instance, closes the
 // transport and persistent storage modules.
 func (nh *NodeHost) Stop() {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		panic(ErrClosed)
+	}
 	atomic.StoreInt32(&nh.closed, 1)
 	nh.sysListener.Publish(server.SystemEvent{
 		Type: server.NodeHostShuttingDown,
@@ -450,7 +452,6 @@ func (nh *NodeHost) StartCluster(initialMembers map[uint64]string,
 	join bool,
 	createStateMachine func(uint64, uint64) sm.IStateMachine,
 	config config.Config) error {
-	nh.ensureNotClosed()
 	stopc := make(chan struct{})
 	cf := func(clusterID uint64, nodeID uint64,
 		done <-chan struct{}) rsm.IManagedStateMachine {
@@ -467,7 +468,6 @@ func (nh *NodeHost) StartConcurrentCluster(initialMembers map[uint64]string,
 	join bool,
 	createStateMachine func(uint64, uint64) sm.IConcurrentStateMachine,
 	config config.Config) error {
-	nh.ensureNotClosed()
 	stopc := make(chan struct{})
 	cf := func(clusterID uint64, nodeID uint64,
 		done <-chan struct{}) rsm.IManagedStateMachine {
@@ -484,7 +484,6 @@ func (nh *NodeHost) StartOnDiskCluster(initialMembers map[uint64]string,
 	join bool,
 	createStateMachine func(uint64, uint64) sm.IOnDiskStateMachine,
 	config config.Config) error {
-	nh.ensureNotClosed()
 	stopc := make(chan struct{})
 	cf := func(clusterID uint64, nodeID uint64,
 		done <-chan struct{}) rsm.IManagedStateMachine {
@@ -502,7 +501,9 @@ func (nh *NodeHost) StartOnDiskCluster(initialMembers map[uint64]string,
 // Note that this is not the membership change operation to remove the node
 // from the Raft cluster.
 func (nh *NodeHost) StopCluster(clusterID uint64) error {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return ErrClosed
+	}
 	return nh.stopNode(clusterID, 0, false)
 }
 
@@ -512,7 +513,9 @@ func (nh *NodeHost) StopCluster(clusterID uint64) error {
 // Note that this is not the membership change operation to remove the node
 // from the Raft cluster.
 func (nh *NodeHost) StopNode(clusterID uint64, nodeID uint64) error {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return ErrClosed
+	}
 	return nh.stopNode(clusterID, nodeID, true)
 }
 
@@ -647,7 +650,9 @@ func (nh *NodeHost) GetClusterMembership(ctx context.Context,
 // on local node's knowledge. The returned boolean value indicates whether the
 // leader information is available.
 func (nh *NodeHost) GetLeaderID(clusterID uint64) (uint64, bool, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return 0, false, ErrClosed
+	}
 	v, ok := nh.getCluster(clusterID)
 	if !ok {
 		return 0, false, ErrClusterNotFound
@@ -837,7 +842,9 @@ func (nh *NodeHost) ReadIndex(clusterID uint64,
 // method. See ReadIndex's example for more details.
 func (nh *NodeHost) ReadLocalNode(rs *RequestState,
 	query interface{}) (interface{}, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	rs.mustBeReadyForLocalRead()
 	// translate the rsm.ErrClusterClosed to ErrClusterClosed
 	// internally, the IManagedStateMachine might obtain a RLock before performing
@@ -861,7 +868,9 @@ func (nh *NodeHost) ReadLocalNode(rs *RequestState,
 // implement the statemachine.IExtended interface.
 func (nh *NodeHost) NAReadLocalNode(rs *RequestState,
 	query []byte) ([]byte, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	rs.mustBeReadyForLocalRead()
 	data, err := rs.node.sm.NALookup(query)
 	if err == rsm.ErrClusterClosed {
@@ -879,7 +888,9 @@ var staleReadCalled uint32
 // ReadIndex and ReadLocalNode method to achieve linearizable read.
 func (nh *NodeHost) StaleRead(clusterID uint64,
 	query interface{}) (interface{}, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	if atomic.CompareAndSwapUint32(&staleReadCalled, 0, 1) {
 		plog.Warningf("StaleRead called, linearizability not guaranteed for stale read")
 	}
@@ -972,7 +983,9 @@ func (nh *NodeHost) SyncRequestSnapshot(ctx context.Context,
 // snapshot in the system at the same Raft log index.
 func (nh *NodeHost) RequestSnapshot(clusterID uint64,
 	opt SnapshotOption, timeout time.Duration) (*RequestState, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	v, ok := nh.getCluster(clusterID)
 	if !ok {
 		return nil, ErrClusterNotFound
@@ -998,7 +1011,9 @@ func (nh *NodeHost) RequestSnapshot(clusterID uint64,
 // nothing to be reclaimed.
 func (nh *NodeHost) RequestCompaction(clusterID uint64,
 	nodeID uint64) (*SysOpState, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	v, ok := nh.getCluster(clusterID)
 	if !ok {
 		// assume this is a node that has already been removed via RemoveData
@@ -1117,7 +1132,9 @@ func (nh *NodeHost) SyncRequestAddWitness(ctx context.Context,
 func (nh *NodeHost) RequestDeleteNode(clusterID uint64,
 	nodeID uint64,
 	configChangeIndex uint64, timeout time.Duration) (*RequestState, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	v, ok := nh.getCluster(clusterID)
 	if !ok {
 		return nil, ErrClusterNotFound
@@ -1154,7 +1171,9 @@ func (nh *NodeHost) RequestDeleteNode(clusterID uint64,
 func (nh *NodeHost) RequestAddNode(clusterID uint64,
 	nodeID uint64, address string, configChangeIndex uint64,
 	timeout time.Duration) (*RequestState, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	v, ok := nh.getCluster(clusterID)
 	if !ok {
 		return nil, ErrClusterNotFound
@@ -1191,7 +1210,9 @@ func (nh *NodeHost) RequestAddNode(clusterID uint64,
 func (nh *NodeHost) RequestAddObserver(clusterID uint64,
 	nodeID uint64, address string, configChangeIndex uint64,
 	timeout time.Duration) (*RequestState, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	v, ok := nh.getCluster(clusterID)
 	if !ok {
 		return nil, ErrClusterNotFound
@@ -1226,7 +1247,9 @@ func (nh *NodeHost) RequestAddObserver(clusterID uint64,
 func (nh *NodeHost) RequestAddWitness(clusterID uint64,
 	nodeID uint64, address string, configChangeIndex uint64,
 	timeout time.Duration) (*RequestState, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	v, ok := nh.getCluster(clusterID)
 	if !ok {
 		return nil, ErrClusterNotFound
@@ -1244,7 +1267,9 @@ func (nh *NodeHost) RequestAddWitness(clusterID uint64,
 // fail after a successful return of the RequestLeaderTransfer method.
 func (nh *NodeHost) RequestLeaderTransfer(clusterID uint64,
 	targetNodeID uint64) error {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return ErrClosed
+	}
 	v, ok := nh.getCluster(clusterID)
 	if !ok {
 		return ErrClusterNotFound
@@ -1311,7 +1336,9 @@ func (nh *NodeHost) RemoveData(clusterID uint64, nodeID uint64) error {
 }
 
 func (nh *NodeHost) removeData(clusterID uint64, nodeID uint64) error {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return ErrClosed
+	}
 	if err := nh.logdb.RemoveNodeData(clusterID, nodeID); err != nil {
 		panic(err)
 	}
@@ -1328,7 +1355,9 @@ func (nh *NodeHost) removeData(clusterID uint64, nodeID uint64) error {
 // the NodeHost. A possible use case is when loading a large data set say with
 // billions of proposals into the dragonboat based system.
 func (nh *NodeHost) GetNodeUser(clusterID uint64) (INodeUser, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	v, ok := nh.getCluster(clusterID)
 	if !ok {
 		return nil, ErrClusterNotFound
@@ -1344,7 +1373,9 @@ func (nh *NodeHost) GetNodeUser(clusterID uint64) (INodeUser, error) {
 // HasNodeInfo returns a boolean value indicating whether the specified node
 // has been bootstrapped on the current NodeHost instance.
 func (nh *NodeHost) HasNodeInfo(clusterID uint64, nodeID uint64) bool {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return false
+	}
 	_, err := nh.logdb.GetBootstrapInfo(clusterID, nodeID)
 	if err == raftio.ErrNoBootstrapInfo {
 		return false
@@ -1359,7 +1390,9 @@ func (nh *NodeHost) HasNodeInfo(clusterID uint64, nodeID uint64) bool {
 // of the NodeHost, this includes details of all Raft clusters managed by the
 // the NodeHost instance.
 func (nh *NodeHost) GetNodeHostInfo(opt NodeHostInfoOption) *NodeHostInfo {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return &NodeHostInfo{}
+	}
 	nhi := &NodeHostInfo{
 		RaftAddress:     nh.RaftAddress(),
 		ClusterInfoList: nh.getClusterInfo(),
@@ -1376,7 +1409,9 @@ func (nh *NodeHost) GetNodeHostInfo(opt NodeHostInfoOption) *NodeHostInfo {
 
 func (nh *NodeHost) propose(s *client.Session,
 	cmd []byte, timeout time.Duration) (*RequestState, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, ErrClosed
+	}
 	v, ok := nh.getClusterNotLocked(s.ClusterID)
 	if !ok {
 		return nil, ErrClusterNotFound
@@ -1391,7 +1426,9 @@ func (nh *NodeHost) propose(s *client.Session,
 
 func (nh *NodeHost) readIndex(clusterID uint64,
 	timeout time.Duration) (*RequestState, *node, error) {
-	nh.ensureNotClosed()
+	if atomic.LoadInt32(&nh.closed) != 0 {
+		return nil, nil, ErrClosed
+	}
 	n, ok := nh.getClusterNotLocked(clusterID)
 	if !ok {
 		return nil, nil, ErrClusterNotFound
@@ -1548,7 +1585,7 @@ func (nh *NodeHost) startCluster(initialMembers map[uint64]string,
 	nh.mu.Lock()
 	defer nh.mu.Unlock()
 	if nh.mu.stopped {
-		return ErrSystemStopped
+		return ErrClosed
 	}
 	if _, ok := nh.mu.clusters.Load(clusterID); ok {
 		return ErrClusterAlreadyExist
@@ -1986,12 +2023,6 @@ func (nh *NodeHost) logNodeHostDetails() {
 		plog.Infof("logdb type: %s", nh.logdb.Name())
 	}
 	plog.Infof("nodehost address: %s", nh.nhConfig.RaftAddress)
-}
-
-func (nh *NodeHost) ensureNotClosed() {
-	if atomic.LoadInt32(&nh.closed) != 0 {
-		panic(ErrClosed)
-	}
 }
 
 func checkRequestState(ctx context.Context,
